backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community  
langchain-google-genai
langchain-chroma
chromadb